Replace ntp_offset.sh with SystemClockOffset.py
As of CentOS 7, chronyd is used instead of ntpd to discipline the
system clock, so we need to query chronyd to get the system clock
offset. SystemClockOffset.py prints the system clock offset in seconds,
obtained from chronyd. If chronyd is not running or is not synchronized
to a source, SystemClockOffset.py will exit with an error.