Invert sense of the waveguide pressure sensor line
The digital line for the N2 waveguide pressure sensor now is 1 when
waveguide pressure is low, and 0 when waveguide pressure is high.
Since the PMC730 has pull-up resistors for the digital input lines,
this means that we will now report low waveguide pressure if the signal
to the digital line is disconnected at any point, which is what we want. 
On the transmitter side, the pressure switch at the N2 regulator is 
normally open, and when closed by sufficient pressure it ties the
digital input line to ground.