Stop appending burst data to IwrfPulsePacket
Some time ago, burst data were removed from IwrfPulsePacket into
a new IwrfBurstPacket. When this change was applied in KaMerge,
the append of burst data onto IwrfPulsePacket was mistakenly left
in place. A recent correction to allocate the correct amount of
space for the IwrfPulsePacket was causing the appended burst data
to overwrite the allocated memory, so the useless append has been
removed.