Use fractional PLL mode when necessary
The integer PLL mode is more desirable for lower phase noise (according
to the documentation), but supports a maximum reference clock divisor
of 127. Since we currently want to run with a 100 MHz reference clock
and 100 KHz frequency steps, we would need a PLL divisor of 1000, which
is not supported. We now detect the divisor-too-large case and switch
to the fractional PLL mode, which does not have the divisor limitation.
The cost is slightly higher phase noise.